Group Update Drop etc.
The following groups are now available:
- symmetric
- Z/nZ
- dihedral
- dihedral nd tensors under addition (isomorphic to D_n \times ... \times D_n)

They have composition.
Also added, a new nditeration function, similar to nditerED, but takes long instead of double.
Creaed src/python_utility for general functions needed for python handling. Added a function for reading tuples of nd array size, currently not in use anywhere.

I'm gonna start making these commit messages somewhat readable as I want to eventually add other contributors.

+++ src/abstract/groups/dihedral.hpp
@@ -27,7 +27,7 @@ extern PyMethodDef dihedral_methods[];
 extern PyTypeObject dihedralType;
 
 // Methods
-dihedral* dihedralCInit(uint64_t r, uint64_t n, uint8_t s);
+dihedral* dihedralCInit(int64_t r, int64_t n, int64_t s);
 static PyObject* PyDihedral_add_new(PyObject *Pyself, PyObject *arg);
 
 #include "dihedral.cpp"
